backend/requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.36.16
botocore==1.36.16
ccxt==4.4.57
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cramjam==2.9.1
cryptography==44.0.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.8
fastparquet==2024.11.0
flatbuffers==25.1.24
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
multitasking==0.0.11
namex==0.0.8
numba==0.61.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
optuna==4.2.0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.1
shap==0.46.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
ta-lib @ git+https://github.com/TA-Lib/ta-lib-python.git@1582434101ee5387765c0dee97f5b4a465f78458
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
yarl==1.18.3
yfinance==0.2.54
